requirements.txt
wxpython==4.1.1
pyperclip
requests
python-vlc
psutil
youtube-search-python
NBSapi
pysubs2
youtube-transcript-api